+++ src/vfvm_sparsesystem.jl
@@ -107,6 +107,10 @@ mutable struct SparseSystem{Tv,Ti, Tm} <: AbstractSystem{Tv,Ti, Tm}
     """
     uhash::UInt64
 
+    """
+    Data for allocation check
+    """
+    allocs::Int
     
     SparseSystem{Tv,Ti,Tm}() where {Tv,Ti,Tm} = new()
 end
@@ -133,6 +137,7 @@ function  SparseSystem(grid,physics::Physics; matrixindextype=Int32)
     system.boundary_factors=zeros(Tv,maxspec,num_bfaceregions(grid))
     system.species_homogeneous=false
     system.uhash=0x0
+    system.allocs=-1000
     system.factorization=nothing
     return system
 end